requirements.txt
amqp==5.3.1
asgiref==3.8.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7.1
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.16.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
locust==2.37.6
locust-cloud==1.21.9
MarkupSafe==3.0.2
msgpack==1.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
python-dateutil==2.9.0.post0
python-engineio==4.12.1
python-socketio==5.13.0
pyzmq==26.4.0
redis==6.2.0
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
